requirements.txt
firebase-admin
paddleocr
opencv-python
pystray
Pillow
numpy==1.23.4
protobuf==3.20.2